requirements.txt
asgiref==3.11.1
asttokens==3.0.1
attrs==25.4.0
decorator==5.2.1
Django==6.0.2
django-extensions==4.1
django-filter==25.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
executing==2.2.1
inflection==0.5.1
iniconfig==2.3.0
ipython==9.10.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
matplotlib-inline==0.2.1
packaging==26.0
parso==0.8.6
pexpect==4.9.0
pluggy==1.6.0
prompt_toolkit==3.0.52
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.2
PyJWT==2.11.0
pytest==9.0.2
pytest-django==4.12.0
PyYAML==6.0.3
referencing==0.37.0
rpds-py==0.30.0
sqlparse==0.5.5
stack-data==0.6.3
traitlets==5.14.3
uritemplate==4.2.0
wcwidth==0.6.0
